# api.yml
openapi: 3.0.1
info:
  title: ExtranetAPI
  version: '1.0'
paths:
...
  /projects/{projectId}:
    get:
      tags:
        - Project
      summary: Получить проект по идентификатору
      parameters:
        - name: projectId
          in: path
          description: ''
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Получить проект по идентификатору
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProjectDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ProjectDto'
    delete:
      tags:
        - Project
      summary: Удалить проект
      parameters:
        - name: projectId
          in: path
          description: ''
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Удалить проект
          content:
            text/plain:
              schema:
                type: integer
                format: int32
            application/json:
              schema:
                type: integer
                format: int32
            text/json:
              schema:
                type: integer
                format: int32
...
  /project-stages/{projectId}:
    get:
      tags:
        - ProjectStage
      summary: Получить этапы проекта
      parameters:
        - name: projectId
          in: path
          description: ''
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Получить этапы проекта
          content:
            text/plain:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProjectStage'
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProjectStage'
            text/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProjectStage'
  /project-stages/{projectId}/current:
    get:
      tags:
        - ProjectStage
      summary: "Получить текущий этап по проекту\r\nВозвращаем первый следующий после последнего выполненного этапа, если выполненны все, то возваращаем последний"
      parameters:
        - name: projectId
          in: path
          description: ''
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Получить текущий этап по проекту
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProjectStage'
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectStage'
            text/json:
              schema:
                $ref: '#/components/schemas/ProjectStage'
  /project-stages/{projectStageId}/project-stage-updating:
    post:
      tags:
        - ProjectStage
      summary: Обновить основную информацию по этапу проекта
      parameters:
        - name: projectStageId
          in: path
          description: ''
          required: true
          schema:
            type: integer
            format: int32
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectStage'
          text/json:
            schema:
              $ref: '#/components/schemas/ProjectStage'
          application/*+json:
            schema:
              $ref: '#/components/schemas/ProjectStage'
        required: true
      responses:
        '200':
          description: Обновить основную информацию по этапу проекта
          content:
            text/plain:
              schema:
                type: integer
                format: int32
            application/json:
              schema:
                type: integer
                format: int32
            text/json:
              schema:
                type: integer
                format: int32
...
components:
  schemas:
...
    ProjectDto:
      type: object
      properties:
        id:
          type: integer
          format: int32
        projectType:
          type: string
          nullable: true
        contractNumber:
          type: string
          nullable: true
        dateOfStart:
          type: string
          format: date-time
          nullable: true
        dateOfApplication:
          type: string
          format: date-time
        deadLine:
          type: string
          format: date-time
          nullable: true
        endDate:
          type: string
          format: date-time
          nullable: true
        clientId:
          type: integer
          format: int32
        address:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        isCompleted:
          type: boolean
      additionalProperties: false
    ProjectStage:
      type: object
      properties:
        id:
          type: integer
          format: int32
        projectId:
          type: integer
          format: int32
        projectStageCode:
          type: string
          nullable: true
        name:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        completedOn:
          type: string
          format: date-time
          nullable: true
        isCompleted:
          type: boolean
      additionalProperties: false
...